refactor: transfer event subscription to subscribers folder
